requirements.txt
langchain==0.3.12
langchain-core==0.3.40
langchain-community==0.3.12
langchain-google-genai==2.0.5
google-generativeai==0.8.3
faiss-cpu==1.13.2
sentence-transformers==2.7.0
streamlit==1.56.0
streamlit-pdf-viewer==0.0.13
pypdf==6.9.2
python-dotenv==1.2.2